Added test to ui-logs controller
